--- app/javascript/packs/editor.tsx
@@ -1,3 +1,0 @@
-// import AceEditor from 'react-ace'
-// import ReactAce from 'react-ace/lib/ace'
-// import 'ace-builds/webpack-resolver'
