+++ app/javascript/utils/react-bootloader.jsx
