+++ app/views/layouts/application.html.haml
@@ -7,7 +7,6 @@
     = csp_meta_tag
     = action_cable_meta_tag
     = javascript_pack_tag 'application', 'data-turbolinks-track': 'reload'
-    = javascript_pack_tag 'editor', 'data-turbolinks-track': 'reload'
     = stylesheet_pack_tag 'application', media: 'all', 'data-turbolinks-track': 'reload'
 
     %meta{ name: "language-server-url", content: Exercism.config.language_server_url }
